fix: UI updates, Added event / volunteer matching
Updated css to match new theme for the site on admin pages.

Updated how some active cards are displayed. Updated buttons and select options in admin pages to appear more modern.

Event / volunteer matching functionality. Selecting an event on the left side of the screen filters out the volunteer options on the right side of the screen according to how ever many skills are desired to match (i.e. if the admin wants the volunteer to have all skills specified by the event, they can specify as such, or they can choose to be more lenient and filter down to allow those who have at least one of the listed skills (even 0 skills), and anywhere in between).