feat: added page for event creation / customization
Page created on admin site to create or modify events.

On modify events, currently there exists a table for all events.

I have a skeleton search bar that will be be utilized to filter the search results, and the intention is that a selected event would take you to modify its details.