fix: finalized validation for event form creation / updates
Event creation form has better validation and is more current in showing notifications to the user as to what is wrong with their inputs.

When a user has completed, they will now see a modal window showing all details for the event, and will have the option to edit their changes, or to accept them.

Accepting them will lead to sending the form data (already handled) to the database (needs to be done).