fix: Volunteer Matching now connects to the Database
The user can now add volunteers to an event and send data to the DB. If the user selects an event, first all of that event's registered volunteers are retrieved from the database and on the frontend, all those volunteer cards are seen as red (you can't add a volunteer to an event more than once).

If the user previously selected a volunteer and then clicked to an event they are registered to, that volunteer is removed from their selection as further protection.