fix: major bugs dealt with on Volunteer Matching
There were adverse effects when using the slider option where selecting a card post filter would select the card at the position the volunteer you picked was originally one (example: volunteers A,B,C,D,E at indices 0,1,2,3,4. Filtering out C would leave you with A,B,D at indices 0,1,2,3. Selecting D would actually select the card at index 4 since D was originally at that index).

Above was fixed, and added some QOL improvements such as resetting the page to the first when adjusting the slider, or selecting a new event. Also, when doing these movements, I implemented the removal of previosuly selected volunteers to avoid potential issues in selecting volunteers and then going to an event they arent allowed to be added. They wouldnt have been selectable, but if they previously were and then the admin tried to send it to the DB, it could be a problem. Though no longer a potential issue now.

The only thing I think I could update to make this page better is to add a notification to display to the user when their filters would yield no volunteers, but they can see it on the right side. That and perhaps better instructions at the top.